fix something

+++ app/src/main/java/com/example/ecommercemobileapp2hand/Models/UserOrder.java
@@ -1,17 +1,18 @@
 package com.example.ecommercemobileapp2hand.Models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 
 public class UserOrder {
     private int user_order_id;
     private int user_id;
-    private double total_price;
+    private BigDecimal total_price;
     private String order_status_id;
     private LocalDateTime created_at;
 
     public UserOrder(){}
-    public UserOrder(int user_order_id, int user_id, double total_price, String order_status_id, LocalDateTime created_at) {
+    public UserOrder(int user_order_id, int user_id, BigDecimal total_price, String order_status_id, LocalDateTime created_at) {
         this.user_order_id = user_order_id;
         this.user_id = user_id;
         this.total_price = total_price;
@@ -35,11 +36,11 @@ public class UserOrder {
         this.user_id = user_id;
     }
 
-    public double getTotal_price() {
+    public BigDecimal getTotal_price() {
         return total_price;
     }
 
-    public void setTotal_price(double total_price) {
+    public void setTotal_price(BigDecimal total_price) {
         this.total_price = total_price;
     }
